hw: save registers for non-idle threads only
When a thread vanishs by a remote action, actually only possible
by using `pause_thread()`, the current thread on that cpu will
become "idle", even if the effected cpu enters the kernel after
receiving an IPI and saves the register of the paused thread.

This commit is no solution for the shortcoming of `pause_thread`
not being synchronized with the calling cpu, but it creates a
general line of defence to save the integrity of the idle threads,
which never need to save their non-changing register set.

Issue genodelabs/genode#5719